requirements.txt
streamlit
jamaibase
python-dotenv
pydantic
pydantic-settings
pytest
